+++ Editor/Authoring/Core/Deployment/IEnvironmentProvider.cs
@@ -1,4 +1,4 @@
-﻿namespace Unity.Services.CloudCode.Authoring.Editor.Core.Deployment
+namespace Unity.Services.CloudCode.Authoring.Editor.Core.Deployment
 {
     interface IEnvironmentProvider
     {