+++ Editor/Authoring/Core/Deployment/ModuleGeneration/Exceptions/FailedProjectRetrieverException.cs
@@ -1,0 +1,12 @@
+using System;
+
+namespace Unity.Services.CloudCode.Authoring.Editor.Core.Deployment.ModuleGeneration.Exceptions
+{
+    class FailedProjectRetrieverException : Exception
+    {
+        public FailedProjectRetrieverException(string message)
+            : base(message)
+        {
+        }
+    }
+}