+++ Editor/Authoring/Core/Deployment/ModuleGeneration/Exceptions/FailedZipCompilationException.cs
@@ -1,0 +1,12 @@
+using System;
+
+namespace Unity.Services.CloudCode.Authoring.Editor.Core.Deployment.ModuleGeneration.Exceptions
+{
+    class FailedZipCompilationException : Exception
+    {
+        public FailedZipCompilationException(Exception ex)
+            : base(ex.Message, ex)
+        {
+        }
+    }
+}