+++ Editor/Authoring/Core/Deployment/ModuleZipper.cs
@@ -1,9 +1,9 @@
-using System;
 using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
 using Unity.Services.CloudCode.Authoring.Editor.Core.Deployment.ModuleGeneration.Exceptions;
 using Unity.Services.CloudCode.Authoring.Editor.Core.IO;
+using Unity.Services.CloudCode.Authoring.Editor.Core.Logging;
 
 namespace Unity.Services.CloudCode.Authoring.Editor.Core.Deployment
 {
@@ -12,10 +12,12 @@ namespace Unity.Services.CloudCode.Authoring.Editor.Core.Deployment
         const string k_ZipFileExtension = "ccm";
 
         readonly IFileSystem m_FileSystem;
+        readonly ILogger m_Logger;
 
-        public ModuleZipper(IFileSystem fileSystem)
+        public ModuleZipper(IFileSystem fileSystem, ILogger logger)
         {
             m_FileSystem = fileSystem;
+            m_Logger = logger;
         }
 
         public async Task<string> ZipCompilation(
@@ -31,6 +33,7 @@ namespace Unity.Services.CloudCode.Authoring.Editor.Core.Deployment
 
                 var zippedFileName = Path.ChangeExtension(moduleName, k_ZipFileExtension);
                 var dstFileFullPath = Path.Join(dstPath, zippedFileName);
+                m_Logger.LogVerbose($"Zipping from '{srcPath}' to '{dstFileFullPath}'");
                 // Remove previously generated module
                 if (m_FileSystem.FileExists(dstFileFullPath))
                 {
@@ -42,6 +45,7 @@ namespace Unity.Services.CloudCode.Authoring.Editor.Core.Deployment
             }
             catch (IOException e)
             {
+                m_Logger.LogVerbose($"Fail to zip {e}");
                 throw new FailedZipCompilationException(e);
             }
         }