+++ Editor/Authoring/Core/Model/StatusSeverityLevel.cs
@@ -1,4 +1,4 @@
-﻿namespace Unity.Services.CloudCode.Authoring.Editor.Core.Model
+namespace Unity.Services.CloudCode.Authoring.Editor.Core.Model
 {
     enum StatusSeverityLevel
     {