+++ Editor/Authoring/Deployment/Modules/TemplateInfo.cs
@@ -3,7 +3,7 @@ using Unity.Services.CloudCode.Authoring.Editor.Core.Solution;
 
 namespace Unity.Services.CloudCode.Authoring.Editor.Deployment.Modules
 {
-    class TemplateInfo: ITemplateInfo
+    class TemplateInfo : ITemplateInfo
     {
         public string PathSolution => Path.Combine(CloudCodePackage.EditorPath,
             "Authoring/Core/Solution/Template~/Solution.sln");
@@ -12,8 +12,8 @@ namespace Unity.Services.CloudCode.Authoring.Editor.Deployment.Modules
         public string PathExampleClass => Path.Combine(CloudCodePackage.EditorPath,
             "Authoring/Core/Solution/Template~/Project/Example.cs");
         public string PathConfig => Path.Combine(CloudCodePackage.EditorPath,
-                "Authoring/Core/Solution/Template~/Project/Properties/PublishProfiles/FolderProfile.pubxml");
+            "Authoring/Core/Solution/Template~/Project/Properties/PublishProfiles/FolderProfile.pubxml");
         public string PathConfigUser => Path.Combine(CloudCodePackage.EditorPath,
-                "Authoring/Core/Solution/Template~/Project/Properties/PublishProfiles/FolderProfile.pubxml.user");
+            "Authoring/Core/Solution/Template~/Project/Properties/PublishProfiles/FolderProfile.pubxml.user");
     }
 }