+++ Runtime/com.unity.services.cloudcode.internal/Http/IsolatedJsonConvert.cs
@@ -11,7 +11,7 @@ namespace Unity.Services.CloudCode.Internal.Http
     {
         [DebuggerStepThrough]
         public static string SerializeObject(object value) => SerializeObject(value, null, null);
-        
+
         [DebuggerStepThrough]
         public static string SerializeObject(object value, JsonSerializerSettings settings) => SerializeObject(value, null, settings);
 
@@ -36,7 +36,7 @@ namespace Unity.Services.CloudCode.Internal.Http
 
             return stringWriter.ToString();
         }
-        
+
         [DebuggerStepThrough]
         public static object DeserializeObject(string value, Type type) => DeserializeObject(value, type, null);